refactor activeTab to tab
