+++ src/tab/analytics/analytics.js
