+++ src/tab/formatter/formatText.js
