+++ src/tab/triggers.js
