+++ src/tab/uiEnhancer/phind/historySearch/insertHistoryBar.js
