+++ src/tab/uiEnhancer/phind/improvePhindUI.js
