+++ src/tab/uiEnhancer/uiEnhancer.js
