+++ src/tab/utils/detectClicksModal.js
